+++ SandboxiePlus/SandMan/Views/TraceView.cpp
@@ -399,11 +399,11 @@ void CTraceView::Refresh()
 		else
 		{
 			if (bHasFilter && !m_pTrace->m_bHighLight) {
-				if (!pEntry->GetName().contains(m_pTrace->m_FilterExp)
-					&& !pEntry->GetMessage().contains(m_pTrace->m_FilterExp)
-					//&& !pEntry->GetTypeStr().contains(m_pTrace->m_FilterExp) // dont filter on non static strings !!!
-					//&& !pEntry->GetStautsStr().contains(m_pTrace->m_FilterExp) // dont filter on non static strings !!!
-					&& !pEntry->GetProcessName().contains(m_pTrace->m_FilterExp))
+				if (!pEntry->GetName().contains(m_pTrace->m_FilterExp, Qt::CaseInsensitive)
+					&& !pEntry->GetMessage().contains(m_pTrace->m_FilterExp, Qt::CaseInsensitive)
+					//&& !pEntry->GetTypeStr().contains(m_pTrace->m_FilterExp, Qt::CaseInsensitive) // dont filter on non static strings !!!
+					//&& !pEntry->GetStautsStr().contains(m_pTrace->m_FilterExp, Qt::CaseInsensitive) // dont filter on non static strings !!!
+					&& !pEntry->GetProcessName().contains(m_pTrace->m_FilterExp, Qt::CaseInsensitive))
 						continue;
 			}